requirements.txt
# MCP Protocol Dependencies
mcp>=1.0.0

# Bitrix24 API Client
fast-bitrix24>=1.5.0

# Async HTTP Client
aiohttp>=3.8.0

# Configuration Management
python-dotenv>=1.0.0

# Type Checking and Validation
beartype>=0.15.0
icontract>=2.6.0
pydantic>=2.5.0

# Progress Visualization
tqdm>=4.64.0

# Development Dependencies
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# pytest-cov>=4.0.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
